requirements.txt
ofxparse==0.21
pandas==2.2.2
python-dotenv==1.0.1
langchain_community==0.3.26
langchain-ollama==0.3.3
crewai==0.165.1
crewai-tools==0.62.3
websocket==0.2.1
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
pymysql==1.1.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
numpy==1.24.3
pydantic==2.5.0
pydantic[email]==2.5.0
python-dotenv==1.0.0
openpyxl==3.1.2
aiofiles==23.2.1
jinja2==3.1.2